requirements.txt
absl-py==0.10.0
mock==4.0.2
prometheus_client==0.8.0
speedtest-cli==2.1.3
requests==2.25.1
